# Make accessing Bigarray.t faster by optimizing Cmmgen.bigarray_indexing