… testsuite/tests/lib-bigarray/bigarrays.reference