I've made Bigarrays get indexed the same way the ordinary arrays are.  This allows the compiler to simplify the expression that accesses the element of bigarrays.  This benchmark shows a noticeable speedup.  Especially for float64 arrays it's in the range of 47-73% on Sandy Bridge CPU:

https://github.com/vbrankov/ocaml-demos/blob/master/optimize-bigarray-indexing/benchmark.ml

 |Nehalem old|Nehalem new|Sandy Bridge old|Sandy Bridge new
------|------|-----|-----|----|
float32 get | 0.615719 | 0.351901 | 0.495019 | 0.324152
float32 set | 0.664388 | 0.460025 | 0.735373 | 0.489247
float64 get | 0.409994 | 0.33201 | 0.421405 | 0.243133
float64 set | 0.525567 | 0.354141 | 0.590739 | 0.400817
int8_signed get | 0.82061 | 0.595768 | 0.712532 | 0.489143
int8_signed set | 0.60036 | 0.366236 | 0.663731 | 0.38448
int8_unsigned get | 0.822517 | 0.595183 | 0.712197 | 0.493002
int8_unsigned set | 0.600529 | 0.36784 | 0.663923 | 0.384583
int16_signed get | 0.824838 | 0.583832 | 0.714127 | 0.447018
int16_signed set | 0.61372 | 0.338638 | 0.666231 | 0.376622
int16_unsigned get | 0.822557 | 0.587342 | 0.753349 | 0.447307
int16_unsigned set | 0.620892 | 0.346735 | 0.666593 | 0.37689
int32 get | 1.898004 | 1.896556 | 2.085518 | 2.087531
int32 set | 0.632254 | 0.34337 | 0.672314 | 0.383324
int64 get | 1.893416 | 1.892835 |2.085168 | 2.087272
int64 set | 0.543767 | 0.382 | 0.592302 | 0.401396
int get | 0.830689 | 0.635045 | 0.798619 | 0.558872
int set | 0.535495 | 0.354656 | 0.563422 | 0.401717
nativeint get | 1.896705 | 1.895186 | 2.086356 | 2.087844
nativeint set | 0.511256 | 0.354471 | 0.559552 | 0.400525
Array2 float64 get | 0.42363 | 0.364187 | 0.356808 | 0.291749
Array2 float64 set | 0.689077 | 0.659695 | 0.754017 | 0.621431
Array3 float64 get | 0.470686 | 0.410212 | 0.429545 | 0.324134
Array3 float64 set | 0.952252 | 1.024793 | 1.010888 | 1.098284