@hcarty 2015-02-15
I haven't had a chance to test this yet but it looks like a very nice addition

What is the cause of the opam error in the automatic test?


`[ERROR] Invalid_argument("'+' is not allowed in compiler versions")`