@vbrankov 2015-02-17
Hezekiah, can you give me a little more information on how do you run the opam test and why do you think the error is linked to this patch.  Thanks.