@dbuenzli 2015-02-17
Seems related to https://github.com/ocaml/opam/pull/1999