@hcarty 2015-02-17
@vbrankov The testing I was referring to is the automatic Travis CI testing run against pull requests.  See the red "Failed" block.  I haven't attempted to run any tests by hand.

@dbuenzli is probably correct in identifying the source of the issue.