@gasche 2015-04-12
This patch needs to either get a review from a specialist, or get written in a style that makes it readable by a non-specialist (for example with comments that explain the dark magic integer tagging tricks). I have no idea why `check_bound bound (untag_int idx) (add_int (mul_int rem bound) (untag_int idx))` would be equivalent to `check_bound bound (untag_int idx) (add_int (mul_int (decr_int rem) bound) idx)`, and would expect a comment to explain what is going on.

(It seems that it also includes a change to make bigarray access respect `-unsafe` when it previously didn't. I find this change itching, given that we have rather been moving away from `-unsafe` as a global setting to more local choices of which primitive to use.)