@vbrankov 2015-04-23
@gasche I've added more comments to make it clearer.  There shouldn't be any changes regarding unsafe, you may want to examine ```check_bound```.  The unsafe in Bigarray lambda is not the same as the option ```-unsafe```.  Understanding the Bigarray handling of ```Translcore``` and ```Cmmgen``` might be necessary to review this patch.