@gasche 2015-04-24
Thanks for the extra comments. I find it very useful to have comments telling whether any given expression is expected to represent an OCaml tagged integer, and which represent a native, untagged integer -- this is one of the most perplexing aspect when reading code at this level. (I find "tagged" and "untagged" easier to follow than "int" and "nativeint", but maybe that's an existing convention in this part of the codebase.)

There are still things I do not understand in this code. Reading the patch with little knowledge of the context, here is what I see:

1. `check_bound` is unchanged
2. however, `check_bound` call-site in the easy case on line 796 is changed, with only the first index-like argument untagged (the second isn't anymore)
3. `let bound = ...` is unchanged
4. again, the call on lines 810-811 only untags the first argument.
5. the call also does something weird with the second one, turning `rem` into `decr_int rem` for no discernable purpose

My assumption, that explains (2) and (4), is that the specification of the `ba_indexing` function changed, it now returns a tagged integer instead of an untagged integer. While the comment before `ba_indexing` in fact says that the function returns a tagged integer (but it took me a few re-reading to understand that it said that), the fact that this is a change is not at all clear. Of course, comments should in general not discuss the previous state of the code; this is what the commit message is for.

That still doesn't explain much of (5). I guess that you use the fact that tagging/untagging through a multiplication cancels out to `decr_int`, but this part specifically is too tricky to review for someone not familiar (or interested) in bit-level tricks without a comment pointing out which equation it is.

Finally, here are two small changes that I think would make the whole thing slightly easier to follow:
- a comment explaining the input/output tagedness of `array_indexing` and `bigarray_indexing`
- renaming the `check_bound` local function to `check_ba_bound`, to make grepping for its use and definition easier to not confuse with the other (semantically different) `check_bound` function in this module.