@vbrankov 2015-04-24
@gasche 

- I meant the local ```check_bound``` was changed.  It's renamed now.
- A part of the speedup comes from simplified tagging.
- I've added a comment to explain ```rem```.  It's necessarily to know the tricks in arithmetic manipulation of tagged integers.  For instance tagged int ```a * b + c``` is untagged as ```((a asr 1) * (b asr 1) + (c asr 1)) lsl 1 + 1```, but ```(a - 1) * b + c``` is simpler and equivalent.