@gasche 2015-04-24
Thanks, I like the patch now.

(For nitpicking only, I think the `(* Untag the index to compare it with the bound *)` comments are useless and could be removed. I suspect the comment on `if unsafe` are also rather transparent, so you should feel free to remove them as well.)