@gasche 2015-04-26
Merged in trunk, thanks.